feat: n5hdf5 reader implements GsonAttributesParser (#11)
* feat: n5hdf5 reader implements GsonAttributesParser

This implementation makes a best effort to distinguish plain string attributes
from object attributes but it will always interpret string attributes that are
indeed valid JSON as object attributes (e.g. '{}', '{"number" : 2}', ...), this can
happen accidentally or on purpose and the only viable solution for this is to
store object attributes as a special type and not as strings which we don't
currently do.